// src/types/eloPlayerType.ts
export interface EloPlayerType {
    queueType: string;
    summonerName: string;
    hotStreak: boolean;
    wins: number;
    veteran: boolean;
    losses: number;
    rank: string;
    leagueId: string;
    inactive: boolean;
    freshBlood: boolean;
    tier: string;
    summonerId: string;
    leaguePoints: number;
}